+++ examples/form/src/ModalTest.vue
@@ -3,13 +3,13 @@ import { createModal } from 'phecda-vue'
 import { ElDialog } from 'element-plus'
 import Test from './components/Test.vue'
 import Test2 from './components/Test2.vue'
-const useWrap = createModal(ElDialog)
+const useModal = createModal(ElDialog)
 function modalAppear() {
   if (Math.random() > 0.5)
-    useWrap(Test2, { name: 'fgs' })
+    useModal(Test2, { name: 'fgs' })
 
   else
-    useWrap(Test, { sex: 1 })
+    useModal(Test, { sex: 1 })
 }
 </script>