+++ examples/vue/src/App.vue
@@ -1,9 +1,10 @@
 <script setup lang="ts">
 import { RouterLink, RouterView } from 'vue-router'
-import { clearStorage, useV } from 'phecda-vue'
+import { clearStorage, getTag, useV } from 'phecda-vue'
 import { HomeModel } from './models/home'
 import HelloWorld from './components/HelloWorld.vue'
-const { name, fullName } = useV(HomeModel)
+const { name, fullName, run } = useV(HomeModel)
+console.dir(HomeModel.name)
 </script>
 
 <template>