+++ packages/vue/src/vue/phecda.ts
@@ -1,16 +1,31 @@
 import type { App } from 'vue'
 import { markRaw } from 'vue'
-import { injectProperty } from 'phecda-core'
+import { getTag, injectProperty } from 'phecda-core'
 import { emitter } from '../emitter'
 
 export const phecdaSymbol = Symbol('phecda')
 
-export function createPhecda() {
+export function createPhecda(symbol?: string) {
   const phecda = markRaw({
     install(app: App) {
       app.provide(phecdaSymbol, phecda)
       app.config.globalProperties.$phecda = phecda
+      if (!window._phecda)
+        window._phecda = {}
 
+      if (symbol) {
+        window._phecda[symbol] = {
+          instance: phecda,
+          snapshot: () => {
+            const ret = [] as { key: string; value: any }[]
+            // @ts-expect-error it works
+            for (const [key, value] of phecda.useOMap)
+              ret.push({ key: getTag(key) || key.name, value })
+
+            return ret
+          },
+        }
+      }
       const eventRecord = [] as [string, (event: any) => void][]
       injectProperty('watcher', ({ eventName, instance, key }: { eventName: string; instance: any; key: string }) => {
         const fn = typeof instance[key] === 'function' ? instance[key].bind(instance) : (v: any) => instance[key] = v
@@ -22,11 +37,13 @@ export function createPhecda() {
         eventRecord.forEach(([eventName, handler]) =>
           emitter.off(eventName, handler),
         )
+        if (symbol)
+          delete window._phecda[symbol]
         originUnmount()
       }
     },
     useVMap: new WeakMap(),
-    useOMap: new WeakMap(),
+    useOMap: new (symbol ? Map : WeakMap)(),
     useRMap: new WeakMap(),
     fnMap: new WeakMap(),
     computedMap: new WeakMap(),